% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_docs.R
\name{internal_C}
\alias{blowC}
\alias{checkBoundsCpp}
\alias{cleanStats}
\alias{internal_C}
\alias{order_reg}
\alias{simpleModelsWrap}
\alias{slacf}
\alias{slnlar}
\title{Internal C and C++ function}
\description{
This functions are for internal use only.
}
\author{
Simon Wood and Matteo Fasiolo <matteo.fasiolo@gmail.com>.
}

