% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_synlik.R
\docType{methods}
\name{plot-synlik}
\alias{plot,synlik,missing-method}
\alias{plot-synlik}
\title{Method for plotting an object of class \code{synlik}.}
\usage{
\S4method{plot}{synlik,missing}(x, y, ...)
}
\arguments{
\item{x}{An object of class \code{synlik}.}

\item{y}{Useless argument, only here for compatibility reasons.}

\item{...}{additional arguments to be passed to \code{object@plotFun}.}
}
\description{
It basically calls the slot \code{object@plotFun} with input \code{object@data}, if it has been provided by the user. 
Otherwise it tries to use the \code{plot(x = object@data, y, ...)} generic.
}
\examples{
data(ricker_sl)

# Using ricker_sl@plotFun
plot(ricker_sl)

# Using generic plot, doesn't work well because object@data is a matrix. 
ricker_sl@plotFun <- NULL
plot(ricker_sl)

}
\author{
Matteo Fasiolo <matteo.fasiolo@gmail.com>
}
\seealso{
\code{\link{synlik-class}}, \code{\link{plot}}.
}

